Removing compile warning
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@1586 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ daq/hlt/src/UnitManager.cc
@@ -59,7 +59,7 @@ EStatus UnitManager::broadCasting(void)
     if (returnCode != c_Success)
       return returnCode;
   }
-  for (int i = 0; (int)i < m_workerNodes.size(); i++) {
+  for (unsigned int i = 0; i < m_workerNodes.size(); i++) {
     if (m_workerNodes[i] != NULL) {
       returnCode = m_workerNodes[i]->broadCasting();
       if (returnCode != c_Success)
